requirements/commons.txt
asgiref==3.3.1
Django==3.1.6
model-bakery==1.2.1
pytz==2021.1
sqlparse==0.4.1
requests==2.25.1
celery==5.0.5
django-celery-beat==2.2.0
redis==3.5.3
djangorestframework==3.12.2
beautifulsoup4==4.9.3
